refactor shared components to shared components dir
